Move .NET 4.5 Kinesis sink to a separate namespace
