+++ serilog-sinks-amazonkinesis.sln
@@ -14,11 +14,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "assets", "assets", "{E9D1B5
 		assets\Serilog.snk = assets\Serilog.snk
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.AmazonKinesis", "src\Serilog.Sinks.AmazonKinesis\Serilog.Sinks.AmazonKinesis.csproj", "{8164BDE4-48FE-4A81-B33E-FEE348904BB8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.Amazon.Kinesis.Stream", "src\Serilog.Sinks.Amazon.Kinesis.Stream\Serilog.Sinks.Amazon.Kinesis.Stream.csproj", "{8164BDE4-48FE-4A81-B33E-FEE348904BB8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sample", "sample", "{543E1315-71DE-40CE-A179-4B27B7A64125}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AmazonKinesisSample", "sample\AmazonKinesisSample\AmazonKinesisSample.csproj", "{6672FF6B-5243-4588-986A-EEA9C0BD6C0B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.Amazon.Kinesis", "src\Serilog.Sinks.Amazon.Kinesis\Serilog.Sinks.Amazon.Kinesis.csproj", "{9CDDC147-93BB-47DC-899C-B41384D7AE23}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,6 +36,10 @@ Global
 		{6672FF6B-5243-4588-986A-EEA9C0BD6C0B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6672FF6B-5243-4588-986A-EEA9C0BD6C0B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6672FF6B-5243-4588-986A-EEA9C0BD6C0B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9CDDC147-93BB-47DC-899C-B41384D7AE23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9CDDC147-93BB-47DC-899C-B41384D7AE23}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9CDDC147-93BB-47DC-899C-B41384D7AE23}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9CDDC147-93BB-47DC-899C-B41384D7AE23}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -41,5 +47,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{8164BDE4-48FE-4A81-B33E-FEE348904BB8} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{6672FF6B-5243-4588-986A-EEA9C0BD6C0B} = {543E1315-71DE-40CE-A179-4B27B7A64125}
+		{9CDDC147-93BB-47DC-899C-B41384D7AE23} = {037440DE-440B-4129-9F7A-09B42D00397E}
 	EndGlobalSection
 EndGlobal
