+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/Properties/AssemblyInfo.cs
