+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/Serilog.Sinks.Amazon.Kinesis.Stream-net40.csproj
