+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/Serilog.Sinks.AmazonKinesis.nuspec
