+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/Sinks/AmazonKinesis/HttpLogShipper-net40.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright 2014 Serilog Contributors
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -22,7 +22,7 @@ using System.Threading;
 using Amazon.Kinesis.Model;
 using Serilog.Debugging;
 
-namespace Serilog.Sinks.AmazonKinesis
+namespace Serilog.Sinks.Amazon.Kinesis.Stream
 {
     class HttpLogShipper : IDisposable
     {
@@ -68,7 +68,7 @@ namespace Serilog.Sinks.AmazonKinesis
                 handler(this, e);
             }
         }
-        
+
         void CloseAndFlush()
         {
             lock (_stateLock)
@@ -113,7 +113,7 @@ namespace Serilog.Sinks.AmazonKinesis
         {
             // Note, called under _stateLock
 
-            _timer.Change(_period, Timeout.InfiniteTimeSpan);
+            _timer.Change(_period, TimeSpan.FromDays(30));
         }
 
         void OnTick()
@@ -185,7 +185,7 @@ namespace Serilog.Sinks.AmazonKinesis
                                         SelfLog.WriteLine("Kinesis failed to index record in stream '{0}'. {1} {2} ", _state.Options.StreamName, record.ErrorCode, record.ErrorMessage);
                                     }
                                     // fire event
-                                    OnLogSendError(new LogSendErrorEventArgs(string.Format("Error writing records to {0} ({1} of {2} records failed)", _state.Options.StreamName, response.FailedRecordCount, count),null));
+                                    OnLogSendError(new LogSendErrorEventArgs(string.Format("Error writing records to {0} ({1} of {2} records failed)", _state.Options.StreamName, response.FailedRecordCount, count), null));
                                 }
                                 else
                                 {
@@ -197,16 +197,15 @@ namespace Serilog.Sinks.AmazonKinesis
                             else
                             {
                                 SelfLog.WriteLine("Found no records to process");
-    
+
                                 // Only advance the bookmark if no other process has the
                                 // current file locked, and its length is as we found it.
 
                                 var bufferedFilesCount = fileSet.Length;
-                                var isProcessingFirstFile = fileSet.First().Equals(currentFilePath,StringComparison.InvariantCultureIgnoreCase);
-                                var isFirstFileUnlocked = IsUnlockedAtLength(currentFilePath, nextLineBeginsAtOffset);
+                                var isProcessingFirstFile = fileSet.First().Equals(currentFilePath, StringComparison.InvariantCultureIgnoreCase);
                                 //SelfLog.WriteLine("BufferedFilesCount: {0}; IsProcessingFirstFile: {1}; IsFirstFileUnlocked: {2}", bufferedFilesCount, isProcessingFirstFile, isFirstFileUnlocked);
 
-                                if (bufferedFilesCount == 2 && isProcessingFirstFile && isFirstFileUnlocked)
+                                if (bufferedFilesCount == 2 && isProcessingFirstFile && IsUnlockedAtLength(currentFilePath, nextLineBeginsAtOffset))
                                 {
                                     SelfLog.WriteLine("Advancing bookmark from '{0}' to '{1}'", currentFilePath, fileSet[1]);
                                     WriteBookmark(bookmark, 0, fileSet[1]);
@@ -225,12 +224,13 @@ namespace Serilog.Sinks.AmazonKinesis
                         }
                     }
                 }
+
                 while (count == _batchPostingLimit);
             }
             catch (Exception ex)
             {
                 SelfLog.WriteLine("Exception while emitting periodic batch from {0}: {1}", this, ex);
-                OnLogSendError(new LogSendErrorEventArgs(string.Format("Error in shipping logs to '{0}' stream)", _state.Options.StreamName),ex));
+                OnLogSendError(new LogSendErrorEventArgs(string.Format("Error in shipping logs to '{0}' stream)", _state.Options.StreamName), ex));
             }
             finally
             {
@@ -254,9 +254,18 @@ namespace Serilog.Sinks.AmazonKinesis
             catch (IOException ex)
             {
                 var errorCode = Marshal.GetHRForException(ex) & ((1 << 16) - 1);
-                if (errorCode != 32 && errorCode != 33)
+
+                if (errorCode == 32)
+                {
+                    SelfLog.WriteLine("Log file {0} is locked by another process, bookmark is not advanced: {1}", file, ex);
+                }
+                else if (errorCode == 33)
                 {
                     SelfLog.WriteLine("Unexpected I/O exception while testing locked status of {0}: {1}", file, ex);
+                }
+                else
+                {
+                    throw;
                 }
             }
             catch (Exception ex)
@@ -269,14 +278,14 @@ namespace Serilog.Sinks.AmazonKinesis
 
         static void WriteBookmark(FileStream bookmark, long nextLineBeginsAtOffset, string currentFile)
         {
-            using (var writer = new StreamWriter(bookmark))
-            {
-                writer.WriteLine("{0}:::{1}", nextLineBeginsAtOffset, currentFile);
-            }
+            // Important not to dispose this StreamReader as the stream must remain open.
+            var writer = new StreamWriter(bookmark);
+            writer.WriteLine("{0}:::{1}", nextLineBeginsAtOffset, currentFile);
+            writer.Flush();
         }
 
         // It would be ideal to chomp whitespace here, but not required.
-        static bool TryReadLine(Stream current, ref long nextStart, out string nextLine)
+        static bool TryReadLine(System.IO.Stream current, ref long nextStart, out string nextLine)
         {
             var includesBom = nextStart == 0;
 
@@ -288,10 +297,9 @@ namespace Serilog.Sinks.AmazonKinesis
 
             current.Position = nextStart;
 
-            using (var reader = new StreamReader(current, Encoding.UTF8, false, 128, true))
-            {
-                nextLine = reader.ReadLine();
-            }
+            // Important not to dispose this StreamReader as the stream must remain open.
+            var reader = new StreamReader(current, Encoding.UTF8, false, 128);
+            nextLine = reader.ReadLine();
 
             if (nextLine == null)
                 return false;
@@ -303,7 +311,7 @@ namespace Serilog.Sinks.AmazonKinesis
             return true;
         }
 
-        static void TryReadBookmark(Stream bookmark, out long nextLineBeginsAtOffset, out string currentFile)
+        static void TryReadBookmark(System.IO.Stream bookmark, out long nextLineBeginsAtOffset, out string currentFile)
         {
             nextLineBeginsAtOffset = 0;
             currentFile = null;
@@ -311,10 +319,9 @@ namespace Serilog.Sinks.AmazonKinesis
             if (bookmark.Length != 0)
             {
                 string current;
-                using (var reader = new StreamReader(bookmark, Encoding.UTF8, false, 128, true))
-                {
-                    current = reader.ReadLine();
-                }
+                // Important not to dispose this StreamReader as the stream must remain open.
+                var reader = new StreamReader(bookmark, Encoding.UTF8, false, 128);
+                current = reader.ReadLine();
 
                 if (current != null)
                 {
