+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/Sinks/AmazonKinesis/HttpLogShipper.cs
@@ -14,15 +14,15 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
 using Amazon.Kinesis.Model;
 using Serilog.Debugging;
+using IO = System.IO;
 
-namespace Serilog.Sinks.Amazon.Kinesis.Streams
+namespace Serilog.Sinks.Amazon.Kinesis.Stream
 {
     class HttpLogShipper : IDisposable
     {
@@ -43,9 +43,9 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
             _state = state;
             _period = _state.Options.BufferLogShippingInterval ?? TimeSpan.FromSeconds(5);
             _batchPostingLimit = _state.Options.BatchPostingLimit;
-            _bookmarkFilename = Path.GetFullPath(_state.Options.BufferBaseFilename + ".bookmark");
-            _logFolder = Path.GetDirectoryName(_bookmarkFilename);
-            _candidateSearchPath = Path.GetFileName(_state.Options.BufferBaseFilename) + "*.json";
+            _bookmarkFilename = IO.Path.GetFullPath(_state.Options.BufferBaseFilename + ".bookmark");
+            _logFolder = IO.Path.GetDirectoryName(_bookmarkFilename);
+            _candidateSearchPath = IO.Path.GetFileName(_state.Options.BufferBaseFilename) + "*.json";
 
             _timer = new Timer(s => OnTick());
 
@@ -113,7 +113,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
         {
             // Note, called under _stateLock
 
-            _timer.Change(_period, TimeSpan.FromDays(30));
+            _timer.Change(_period, Timeout.InfiniteTimeSpan);
         }
 
         void OnTick()
@@ -127,7 +127,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
                     // Locking the bookmark ensures that though there may be multiple instances of this
                     // class running, only one will ship logs at a time.
 
-                    using (var bookmark = File.Open(_bookmarkFilename, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read))
+                    using (var bookmark = IO.File.Open(_bookmarkFilename, IO.FileMode.OpenOrCreate, IO.FileAccess.ReadWrite, IO.FileShare.Read))
                     {
                         long startingOffset;
                         long nextLineBeginsAtOffset;
@@ -138,7 +138,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
 
                         var fileSet = GetFileSet();
 
-                        if (currentFilePath == null || !File.Exists(currentFilePath))
+                        if (currentFilePath == null || !IO.File.Exists(currentFilePath))
                         {
                             nextLineBeginsAtOffset = 0;
                             currentFilePath = fileSet.FirstOrDefault();
@@ -149,7 +149,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
                             count = 0;
 
                             var records = new List<PutRecordsRequestEntry>();
-                            using (var current = File.Open(currentFilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
+                            using (var current = IO.File.Open(currentFilePath, IO.FileMode.Open, IO.FileAccess.Read, IO.FileShare.ReadWrite))
                             {
                                 startingOffset = current.Position = nextLineBeginsAtOffset;
 
@@ -161,7 +161,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
                                     var record = new PutRecordsRequestEntry
                                     {
                                         PartitionKey = Guid.NewGuid().ToString(),
-                                        Data = new MemoryStream(bytes)
+                                        Data = new IO.MemoryStream(bytes)
                                     };
                                     records.Add(record);
                                 }
@@ -185,7 +185,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
                                         SelfLog.WriteLine("Kinesis failed to index record in stream '{0}'. {1} {2} ", _state.Options.StreamName, record.ErrorCode, record.ErrorMessage);
                                     }
                                     // fire event
-                                    OnLogSendError(new LogSendErrorEventArgs(string.Format("Error writing records to {0} ({1} of {2} records failed)", _state.Options.StreamName, response.FailedRecordCount, count), null));
+                                    OnLogSendError(new LogSendErrorEventArgs(string.Format("Error writing records to {0} ({1} of {2} records failed)", _state.Options.StreamName, response.FailedRecordCount, count),null));
                                 }
                                 else
                                 {
@@ -202,10 +202,11 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
                                 // current file locked, and its length is as we found it.
 
                                 var bufferedFilesCount = fileSet.Length;
-                                var isProcessingFirstFile = fileSet.First().Equals(currentFilePath, StringComparison.InvariantCultureIgnoreCase);
+                                var isProcessingFirstFile = fileSet.First().Equals(currentFilePath,StringComparison.InvariantCultureIgnoreCase);
+                                var isFirstFileUnlocked = IsUnlockedAtLength(currentFilePath, nextLineBeginsAtOffset);
                                 //SelfLog.WriteLine("BufferedFilesCount: {0}; IsProcessingFirstFile: {1}; IsFirstFileUnlocked: {2}", bufferedFilesCount, isProcessingFirstFile, isFirstFileUnlocked);
 
-                                if (bufferedFilesCount == 2 && isProcessingFirstFile && IsUnlockedAtLength(currentFilePath, nextLineBeginsAtOffset))
+                                if (bufferedFilesCount == 2 && isProcessingFirstFile && isFirstFileUnlocked)
                                 {
                                     SelfLog.WriteLine("Advancing bookmark from '{0}' to '{1}'", currentFilePath, fileSet[1]);
                                     WriteBookmark(bookmark, 0, fileSet[1]);
@@ -218,19 +219,18 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
                                     // will delay this.
                                     SelfLog.WriteLine("Deleting '{0}'", fileSet[0]);
 
-                                    File.Delete(fileSet[0]);
+                                    IO.File.Delete(fileSet[0]);
                                 }
                             }
                         }
                     }
                 }
-
                 while (count == _batchPostingLimit);
             }
             catch (Exception ex)
             {
                 SelfLog.WriteLine("Exception while emitting periodic batch from {0}: {1}", this, ex);
-                OnLogSendError(new LogSendErrorEventArgs(string.Format("Error in shipping logs to '{0}' stream)", _state.Options.StreamName), ex));
+                OnLogSendError(new LogSendErrorEventArgs(string.Format("Error in shipping logs to '{0}' stream)", _state.Options.StreamName),ex));
             }
             finally
             {
@@ -246,26 +246,17 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
         {
             try
             {
-                using (var fileStream = File.Open(file, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read))
+                using (var fileStream = IO.File.Open(file, IO.FileMode.OpenOrCreate, IO.FileAccess.ReadWrite, IO.FileShare.Read))
                 {
                     return fileStream.Length <= maxLen;
                 }
             }
-            catch (IOException ex)
+            catch (IO.IOException ex)
             {
                 var errorCode = Marshal.GetHRForException(ex) & ((1 << 16) - 1);
-
-                if (errorCode == 32)
-                {
-                    SelfLog.WriteLine("Log file {0} is locked by another process, bookmark is not advanced: {1}", file, ex);
-                }
-                else if (errorCode == 33)
+                if (errorCode != 32 && errorCode != 33)
                 {
                     SelfLog.WriteLine("Unexpected I/O exception while testing locked status of {0}: {1}", file, ex);
-                }
-                else
-                {
-                    throw;
                 }
             }
             catch (Exception ex)
@@ -276,16 +267,16 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
             return false;
         }
 
-        static void WriteBookmark(FileStream bookmark, long nextLineBeginsAtOffset, string currentFile)
+        static void WriteBookmark(IO.FileStream bookmark, long nextLineBeginsAtOffset, string currentFile)
         {
-            // Important not to dispose this StreamReader as the stream must remain open.
-            var writer = new StreamWriter(bookmark);
-            writer.WriteLine("{0}:::{1}", nextLineBeginsAtOffset, currentFile);
-            writer.Flush();
+            using (var writer = new IO.StreamWriter(bookmark))
+            {
+                writer.WriteLine("{0}:::{1}", nextLineBeginsAtOffset, currentFile);
+            }
         }
 
         // It would be ideal to chomp whitespace here, but not required.
-        static bool TryReadLine(Stream current, ref long nextStart, out string nextLine)
+        static bool TryReadLine(IO.Stream current, ref long nextStart, out string nextLine)
         {
             var includesBom = nextStart == 0;
 
@@ -297,9 +288,10 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
 
             current.Position = nextStart;
 
-            // Important not to dispose this StreamReader as the stream must remain open.
-            var reader = new StreamReader(current, Encoding.UTF8, false, 128);
-            nextLine = reader.ReadLine();
+            using (var reader = new IO.StreamReader(current, Encoding.UTF8, false, 128, true))
+            {
+                nextLine = reader.ReadLine();
+            }
 
             if (nextLine == null)
                 return false;
@@ -311,7 +303,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
             return true;
         }
 
-        static void TryReadBookmark(Stream bookmark, out long nextLineBeginsAtOffset, out string currentFile)
+        static void TryReadBookmark(IO.Stream bookmark, out long nextLineBeginsAtOffset, out string currentFile)
         {
             nextLineBeginsAtOffset = 0;
             currentFile = null;
@@ -319,9 +311,10 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
             if (bookmark.Length != 0)
             {
                 string current;
-                // Important not to dispose this StreamReader as the stream must remain open.
-                var reader = new StreamReader(bookmark, Encoding.UTF8, false, 128);
-                current = reader.ReadLine();
+                using (var reader = new IO.StreamReader(bookmark, Encoding.UTF8, false, 128, true))
+                {
+                    current = reader.ReadLine();
+                }
 
                 if (current != null)
                 {
@@ -339,7 +332,7 @@ namespace Serilog.Sinks.Amazon.Kinesis.Streams
 
         string[] GetFileSet()
         {
-            return Directory.GetFiles(_logFolder, _candidateSearchPath)
+            return IO.Directory.GetFiles(_logFolder, _candidateSearchPath)
                 .OrderBy(n => n)
                 .ToArray();
         }
