+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/packages.config
